backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-auth==2.29.0
google-api-python-client==2.127.0
python-dotenv==1.0.1
